Add ISP disabler
The idea here is to prevent our long serial cable from loading down SCLK, since the SPI clock is shared with the ISP interface.
However, it turns out that simply adding the 4016 is enough to prevent SCLK from being messed up.
For now we'll still make use of it. It doesn't interfere with ISP