fieldmill8: Mount ITR-20001 in LPPB032NFSP
Pins should fit, but we can file them down a bit if not.
Solder them in place after that.
Another alternative is to solder the ITR-20001's pins to the LPPB032NFSP pads.
The ITR-20001 has been placed at around 3 mm from the rotor,
to give the light a low enough incident angle (too close = blind)